[add]AQI image+[set]layout+[fix]AQI api too many requests
