+++ airviewfrontend/src/index.css
@@ -1,6 +1,7 @@
 body {
   margin: 0;
   font-family: 'Poppins', sans-serif;
+  background-color: rgb(228, 242, 255);
 }
 
 code {
